refactor : s3 테스트 관련 비지니스 코드 이동
